do not use static reference for instances
